Swap constant values of StartOffset
In theory this diff is a no-op because nobody should be providing negative offset integers outside the use of these constants. Unfortunately it turns out there's no real way to make these type safe without making them into structs because constants are untyped until a type is required, meaning that integer literals can be used in place of the iota constant you really wanted to force people to use.