Fix connection pool timeout case
When a broker becomes unavailable and we are failing to connect to it
(dial timeout or connection error) we need to propogate that message
up as the result of GetConnection so that the caller knows something
happened that wasn't a transient issue.

The broken behavior right now doesn't propogate the message which causes
metadata to never be refreshed when a broker goes down.